fix: handle registration 429 errors
